add aux channel
moved application-specific gotRemoteChannel handlers to the application.
It is now easy to add multiple named data channels with their own
handlers.  As an example, the aux channel causes an alert on the remote
client, whereas the whiteboard channel causes a message to post in the
remote's receive textarea (for now; obviously this is intended to be
used by the whiteboard).