modularize signaller as own object (mostly)
There are still some global variables that need to be brought into
object scope, both for the rtc_connection and signaller objects.

Also, the signaller should be able to function optionally without a
messaging pane.

Finally, I still need to finish refactoring the signalling events that
happen when creating an rtc_connection between clients, which is partly
started (extracting signaller was implicated in this...).